Renames exceptions package to exception
