+++ src/main/kotlin/it/unibo/lss/fcla/athleticpreparation/domain/exception/WorkoutPlanException.kt
@@ -1,3 +1,3 @@
-package it.unibo.lss.fcla.athleticpreparation.domain.exceptions
+package it.unibo.lss.fcla.athleticpreparation.domain.exception
 
 class WorkoutPlanException(message: String) : Exception(message)
